# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScanAR API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Model {
  id: string;
  name: string;
  filename: string;
  url: string;
  createdAt: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthUser {
  id: string;
  email: string;
  plan: string;
}

export interface SubscriptionInfo {
  plan: string;
  modelCount: number;
  limit: number;
  canUpload: boolean;
}

export type UpgradePlanRequestPlan =
  (typeof UpgradePlanRequestPlan)[keyof typeof UpgradePlanRequestPlan];

export const UpgradePlanRequestPlan = {
  free: "free",
  pro: "pro",
  business: "business",
} as const;

export interface UpgradePlanRequest {
  plan: UpgradePlanRequestPlan;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}