# lib/api-zod/src/generated/types/subscriptionInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScanAR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionInfo {
  plan: string;
  modelCount: number;
  limit: number;
  canUpload: boolean;
}