# lib/api-zod/src/generated/types/upgradePlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScanAR API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpgradePlanRequestPlan } from "./upgradePlanRequestPlan";

export interface UpgradePlanRequest {
  plan: UpgradePlanRequestPlan;
}